# docs/swagger.yaml
basePath: /v1/api
definitions:
  CONVERDA_internal_apps_controller_dto.APIKeyFullResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      name:
        type: string
      plain_key:
        description: Only returned once on creation/rotation
        type: string
      prefix:
        type: string
      revoked_at:
        type: string
      suffix:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.APIKeyResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      name:
        type: string
      prefix:
        type: string
      revoked_at:
        type: string
      suffix:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.AppResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      environments:
        items:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.EnvironmentResponse'
        type: array
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.CreateAppRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  CONVERDA_internal_apps_controller_dto.CreateEnvironmentRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  CONVERDA_internal_apps_controller_dto.CreateProviderRequest:
    properties:
      configuration:
        additionalProperties: true
        type: object
      environment_id:
        type: string
      provider_name:
        type: string
      provider_type:
        type: string
    required:
    - environment_id
    - provider_name
    - provider_type
    type: object
  CONVERDA_internal_apps_controller_dto.CreateWebhookRequest:
    properties:
      description:
        type: string
      environment_id:
        type: string
      events:
        items:
          type: string
        type: array
      url:
        type: string
    required:
    - environment_id
    - url
    type: object
  CONVERDA_internal_apps_controller_dto.EnvironmentResponse:
    properties:
      app_id:
        type: string
      created_at:
        type: string
      environment_code:
        type: string
      id:
        type: string
      keys:
        items:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.APIKeyResponse'
        type: array
    type: object
  CONVERDA_internal_apps_controller_dto.ProviderResponse:
    properties:
      app_id:
        type: string
      configuration:
        additionalProperties: true
        type: object
      created_at:
        type: string
      environment_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      provider_name:
        type: string
      provider_type:
        type: string
      tenant_id:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.RotateKeyRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  CONVERDA_internal_apps_controller_dto.SystemEnvironmentResponse:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.UpdateAppRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.UpdateProviderRequest:
    properties:
      configuration:
        additionalProperties: true
        type: object
      is_active:
        type: boolean
      provider_name:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.UpdateWebhookRequest:
    properties:
      description:
        type: string
      events:
        items:
          type: string
        type: array
      is_active:
        type: boolean
      url:
        type: string
    type: object
  CONVERDA_internal_apps_controller_dto.WebhookResponse:
    properties:
      app_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      environment_id:
        type: string
      events:
        items:
          type: string
        type: array
      id:
        type: string
      is_active:
        type: boolean
      secret:
        type: string
      tenant_id:
        type: string
      url:
        type: string
    type: object
...
  CONVERDA_internal_iam_controller_dto.CreateTenantRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  CONVERDA_internal_iam_controller_dto.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  CONVERDA_internal_iam_controller_dto.PricingPlanResponse:
    properties:
      currency:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      monthly_credits:
        type: integer
      name:
        type: string
      price:
        type: number
      slug:
        type: string
    type: object
...
  CONVERDA_internal_iam_controller_dto.TenantResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      plan_start_date:
        type: string
      pricing_plan:
        $ref: '#/definitions/CONVERDA_internal_iam_controller_dto.PricingPlanResponse'
      pricing_plan_id:
        type: string
      slug:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a server for a Go Drunk Backend API, demonstrating DDD principles.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Drunk Backend API by DDD
  version: "1.0"
paths:
...
  /api-keys/{key_id}/revoke:
    post:
      parameters:
      - description: Key ID
        in: path
        name: key_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Revoke API Key
      tags:
      - Apps
  /apps/{app_id}:
    delete:
      description: Delete an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete app
      tags:
      - Apps
    get:
      description: Get details of a specific app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.AppResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get app
      tags:
      - Apps
    put:
      consumes:
      - application/json
      description: Update an existing app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.UpdateAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.AppResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update app
      tags:
      - Apps
  /apps/{app_id}/environments:
    get:
      description: List all environments for an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.EnvironmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List environments
      tags:
      - Apps
    post:
      consumes:
      - application/json
      description: Create a new environment for an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      - description: Environment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.CreateEnvironmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.EnvironmentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create environment
      tags:
      - Apps
...
  /apps/{app_id}/providers:
    get:
      description: List all providers for an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.ProviderResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List providers
      tags:
      - Providers
    post:
      consumes:
      - application/json
      description: Create a new provider for an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      - description: Tenant ID
        in: query
        name: tenant_id
        required: true
        type: string
      - description: Provider data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.CreateProviderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a provider
      tags:
      - Providers
  /apps/{app_id}/webhooks:
    get:
      description: List all webhooks for an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.WebhookResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List webhooks
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Create a new webhook for an app
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: string
      - description: Tenant ID
        in: query
        name: tenant_id
        required: true
        type: string
      - description: Webhook data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.CreateWebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.WebhookResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a webhook
      tags:
      - Webhooks
...
  /environments/{id}/api-keys:
    get:
      parameters:
      - description: Env ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.APIKeyResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List API Keys
      tags:
      - Apps
  /environments/{id}/api-keys/rotate:
    post:
      consumes:
      - application/json
      description: Revokes all old keys and generates a new one
      parameters:
      - description: Env ID
        in: path
        name: id
        required: true
        type: string
      - description: Key data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.RotateKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.APIKeyFullResponse'
      security:
      - BearerAuth: []
      summary: Rotate API Key
      tags:
      - Apps
...
  /providers/{provider_id}:
    delete:
      description: Delete a provider
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete provider
      tags:
      - Providers
    get:
      description: Get details of a specific provider
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.ProviderResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get provider
      tags:
      - Providers
    put:
      consumes:
      - application/json
      description: Update an existing provider
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.UpdateProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.ProviderResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update provider
      tags:
      - Providers
...
  /system/environments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.SystemEnvironmentResponse'
            type: array
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: List system environments
      tags:
      - System
  /tenants:
    get:
      description: Get a paginated list of all tenants
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_iam_controller_dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: List all tenants
      tags:
      - Tenants
    post:
      consumes:
      - application/json
      description: Create a new tenant/organization
      parameters:
      - description: Tenant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_iam_controller_dto.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_iam_controller_dto.TenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new tenant
      tags:
      - Tenants
  /tenants/{id}:
    delete:
      description: Delete a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete tenant
      tags:
      - Tenants
    get:
      description: Get a specific tenant by its ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_iam_controller_dto.TenantResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get tenant by ID
      tags:
      - Tenants
    put:
      consumes:
      - application/json
      description: Update an existing tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_iam_controller_dto.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_iam_controller_dto.TenantResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update tenant
      tags:
      - Tenants
  /tenants/{id}/apps:
    get:
      description: List all apps for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.AppResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List apps
      tags:
      - Apps
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: App data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.CreateAppRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.AppResponse'
      security:
      - BearerAuth: []
      summary: Create a new app
      tags:
      - Apps
...
  /webhooks/{webhook_id}:
    delete:
      description: Delete a webhook
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete webhook
      tags:
      - Webhooks
    get:
      description: Get details of a specific webhook
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.WebhookResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get webhook
      tags:
      - Webhooks
    put:
      consumes:
      - application/json
      description: Update an existing webhook
      parameters:
      - description: Webhook ID
        in: path
        name: webhook_id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.UpdateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_apps_controller_dto.WebhookResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update webhook
      tags:
      - Webhooks
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"